修复编译警告作为错误的完整实现
主要更改：
- 添加NPP_WARNINGS_AS_ERRORS编译选项，默认启用
- 修复所有CUDA核心文件中的编译警告：
  * 删除未使用变量：nppi_filter.cu, nppi_warp_affine.cu, nppi_watershed.cu
  * 修复类型转换问题：nppi_magnitude.cpp, nppi_morphology.cu
  * 添加模板特化处理浮点类型：nppi_morphology.cu
  * 修复未使用参数：多个.cpp文件
- 删除过时备份文件：nppi_filterGauss_backup.cu
- 部分修复测试文件警告：test_nppi_absdiff.cpp, test_nppi_comparec.cpp, test_nppi_logical.cpp

✅ 核心NPP库现在可以在警告作为错误模式下完整编译
✅ 基础功能测试通过 (basic_add_example运行成功)

+++ src/nppi/nppi_linear_transform_operations/nppi_magnitude.cpp
@@ -104,7 +104,7 @@ NppStatus nppiMagnitude_32f_C2R(const Npp32f * pSrc, int nSrcStep,
     if (oSizeROI.width <= 0 || oSizeROI.height <= 0) {
         return NPP_SIZE_ERROR;
     }
-    if (nSrcStep < oSizeROI.width * 2 * sizeof(Npp32f) || nDstStep < oSizeROI.width * sizeof(Npp32f)) {
+    if (nSrcStep < oSizeROI.width * 2 * (int)sizeof(Npp32f) || nDstStep < oSizeROI.width * (int)sizeof(Npp32f)) {
         return NPP_STEP_ERROR;
     }
